backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
sqlalchemy==2.0.43
aiomysql
python-dotenv
huggingface_hub
